# docs/swagger.yaml
definitions:
  controller.CreateRatingsParams:
    properties:
      post_id:
        description: 帖子ID
        example: 1
        type: integer
      rating:
        description: 评分 -5 ~ 5
        example: 5
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - post_id
    - rating
    - user_id
    type: object
  controller.GetPostByIdParams:
    properties:
      post_id:
        description: 发帖ID
        example: 1
        type: integer
    type: object
  controller.GetPostByIdResponse:
    properties:
      post:
        allOf:
        - $ref: '#/definitions/store.Post'
        description: 发帖
    type: object
  controller.GetRatingsAveParams:
    properties:
      post_id:
        description: 帖子ID
        example: 1
        type: integer
    required:
    - post_id
    type: object
  controller.GetRatingsAveResponse:
    properties:
      post_id:
        description: 帖子ID
        example: 1
        type: integer
      rating_ave:
        description: 平均评分
        example: 5
        type: number
      rating_count:
        description: 评分总人数
        example: 1
        type: integer
    type: object
  controller.GetRatingsParams:
    properties:
      post_id:
        description: 帖子ID
        example: 1
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - post_id
    - user_id
    type: object
  controller.GetRatingsResponse:
    properties:
      post_id:
        description: 帖子ID
        example: 1
        type: integer
      rating:
        description: 评分
        example: 5
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    type: object
  controller.GetUserByIDRequest:
    properties:
      id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - id
    type: object
  controller.ListPostsByKeywordParams:
    properties:
      cursor:
        example: 0
        type: integer
      keyword:
        description: 关键字，可选
        example: ""
        type: string
    type: object
  controller.ListPostsParams:
    properties:
      cursor:
        example: 0
        type: integer
    type: object
  controller.PostLikeGetParams:
    properties:
      post_id:
        description: 帖子ID
        example: 1
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - post_id
    - user_id
    type: object
  controller.PostLikeGetResponse:
    properties:
      is_like:
        description: '是否点赞: 0 未点赞, 1 已点赞'
        example: 1
        type: integer
    type: object
  controller.PostLikeUpdateLikedParams:
    properties:
      is_like:
        description: '是否点赞: 0 取消点赞, 1 点赞'
        example: 1
        type: integer
      post_id:
        description: 帖子ID
        example: 1
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - post_id
    - user_id
    type: object
  store.Post:
    properties:
      content:
        example: "1"
        type: string
      created_at:
        type: integer
      id:
        example: 1
        type: integer
      images:
        example:
        - 1.jpg
        - 2.jpg
        items:
          type: string
        type: array
      is_active:
        description: 是否展示 0 不展示 1 展示
        example: 1
        type: integer
      latitude:
        example: 120.123456
        type: number
      like_count:
        example: 0
        type: integer
      longitude:
        example: 30.123456
        type: number
      post_tags:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      updated_at:
        type: integer
      user_id:
        example: 1
        type: integer
    required:
    - content
    - post_tags
    - user_id
    type: object
  store.UserRsp:
    properties:
      avatar:
        example: https://example.com/avatar.png
        type: string
      id:
        example: 1
        type: integer
      user_name:
        example: 春日收藏家
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
  title: --Swagger Example API
  version: "1234.0"
paths:
  /post_like/get:
    post:
      description: 0 未点赞, 1 已点赞
      parameters:
      - description: 参数
        in: body
        name: p
        required: true
        schema:
          $ref: '#/definitions/controller.PostLikeGetParams'
      responses:
        "200":
          description: 返回结果
          schema:
            $ref: '#/definitions/controller.PostLikeGetResponse'
      summary: 获取当前用户对指定帖子的点赞状态
      tags:
      - post_like
  /post_like/update_liked:
    post:
      consumes:
      - application/json
      description: 0 取消点赞, 1 点赞
      parameters:
      - description: 参数
        in: body
        name: p
        required: true
        schema:
          $ref: '#/definitions/controller.PostLikeUpdateLikedParams'
      produces:
      - application/json
      responses: {}
      summary: 更新当前用户对指定帖子的点赞状态
      tags:
      - post_like
  /posts/create:
    post:
      consumes:
      - application/json
      description: 创建发帖
      parameters:
      - description: 发帖信息
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/store.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Post'
      summary: 创建发帖
      tags:
      - posts
  /posts/get_by_id:
    post:
      description: 获取单个发帖
      parameters:
      - description: 获取单个发帖信息
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/controller.GetPostByIdParams'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetPostByIdResponse'
      summary: 获取单个发帖
      tags:
      - posts
  /posts/list:
    post:
      consumes:
      - application/json
      description: 获取所有发帖列表
      parameters:
      - description: 发帖信息
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/controller.ListPostsParams'
      produces:
      - application/json
      responses: {}
      summary: 获取所有发帖列表
      tags:
      - posts
  /posts/list_by_keyword:
    post:
      consumes:
      - application/json
      description: 通过关键字获取发帖列表
      parameters:
      - description: 通过关键字获取发帖列表信息
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/controller.ListPostsByKeywordParams'
      produces:
      - application/json
      responses: {}
      summary: 通过关键字获取发帖列表
      tags:
      - posts
  /ratings/create:
    post:
      consumes:
      - application/json
      description: 用户对帖子评分
      parameters:
      - description: 参数示例
        in: body
        name: p
        required: true
        schema:
          $ref: '#/definitions/controller.CreateRatingsParams'
      produces:
      - application/json
      responses: {}
      summary: 用户对帖子评分
      tags:
      - ratings
  /ratings/get:
    post:
      consumes:
      - application/json
      description: 获取用户对帖子的评分
      parameters:
      - description: 参数示例
        in: body
        name: p
        required: true
        schema:
          $ref: '#/definitions/controller.GetRatingsParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetRatingsResponse'
      summary: 获取用户对帖子的评分
      tags:
      - ratings
  /ratings/get_ave:
    post:
      consumes:
      - application/json
      description: 获取帖子评分平均值
      parameters:
      - description: 参数示例
        in: body
        name: p
        required: true
        schema:
          $ref: '#/definitions/controller.GetRatingsAveParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetRatingsAveResponse'
      summary: 获取帖子评分平均值
      tags:
      - ratings
  /tags/list:
    post:
      consumes:
      - application/json
      description: 获取标签列表
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 获取标签列表
      tags:
      - tags
  /user/get_by_id:
    post:
      consumes:
      - application/json
      description: 获取指定用户
      parameters:
      - description: 获取指定用户
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/controller.GetUserByIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.UserRsp'
      summary: 获取指定用户
      tags:
      - user
  /user/new_random_user:
    post:
      consumes:
      - application/json
      description: 获取随机新用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.UserRsp'
      summary: 获取随机新用户
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"